requirements.txt
#django backend
django
djangorestframework
django-cors-headers
django-allauth
PyJWT




#mostly for chatbot (HTTP METHODS)
fastapi
uvicorn

#for RAG
pandas
langchain
langchain-ollama
langchain-chroma

#for stt (speech to text)
openai-whisper
#torch --index-url https://download.pytorch.org/whl/cpu     (install this sepreraely)

#payments
stripe

#extras
python-multipart
python-dotenv
python-decouple
cryptography


pydantic
pillow


#for blood donation feature
geopy


# npm install (for map) 
leaflet 
react-leaflet

#IN COMMAND PROMPT, global
curl -fsSL https://stripe.com/install.sh | sudo bash    #-> this installs the stripe CLI in your device which we will use later to to get out webhook secret (that we are using in our settings.py
